requirements.txt
websocket-client==1.4.2
rel==0.4.7
